drm: move legacy::Error into common module
